refactor: #90 베이스 기타 pacakge 분리
