+++ src/main/java/com/ajou/hertz/domain/instrument/bass_guitar/dto/response/BassGuitarResponse.java
@@ -1,15 +1,17 @@
-package com.ajou.hertz.domain.instrument.dto.response;
+package com.ajou.hertz.domain.instrument.bass_guitar.dto.response;
 
 import java.util.List;
 
 import com.ajou.hertz.common.dto.response.AddressResponse;
-import com.ajou.hertz.domain.instrument.constant.BassGuitarBrand;
-import com.ajou.hertz.domain.instrument.constant.BassGuitarPickUp;
-import com.ajou.hertz.domain.instrument.constant.BassGuitarPreAmplifier;
+import com.ajou.hertz.domain.instrument.bass_guitar.constant.BassGuitarBrand;
+import com.ajou.hertz.domain.instrument.bass_guitar.constant.BassGuitarPickUp;
+import com.ajou.hertz.domain.instrument.bass_guitar.constant.BassGuitarPreAmplifier;
 import com.ajou.hertz.domain.instrument.constant.GuitarColor;
 import com.ajou.hertz.domain.instrument.constant.InstrumentCategory;
 import com.ajou.hertz.domain.instrument.constant.InstrumentProgressStatus;
-import com.ajou.hertz.domain.instrument.dto.BassGuitarDto;
+import com.ajou.hertz.domain.instrument.bass_guitar.dto.BassGuitarDto;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentImageResponse;
+import com.ajou.hertz.domain.instrument.dto.response.InstrumentResponse;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AccessLevel;
