+++ src/main/java/com/ajou/hertz/domain/instrument/bass_guitar/strategy/BassGuitarCreationStrategy.java
@@ -1,7 +1,8 @@
-package com.ajou.hertz.domain.instrument.strategy;
+package com.ajou.hertz.domain.instrument.bass_guitar.strategy;
 
-import com.ajou.hertz.domain.instrument.dto.request.CreateNewBassGuitarRequest;
-import com.ajou.hertz.domain.instrument.entity.BassGuitar;
+import com.ajou.hertz.domain.instrument.bass_guitar.dto.request.CreateNewBassGuitarRequest;
+import com.ajou.hertz.domain.instrument.bass_guitar.entity.BassGuitar;
+import com.ajou.hertz.domain.instrument.strategy.InstrumentCreationStrategy;
 import com.ajou.hertz.domain.user.entity.User;
 
 public class BassGuitarCreationStrategy implements InstrumentCreationStrategy<BassGuitar, CreateNewBassGuitarRequest> {
