+++ src/main/java/com/ajou/hertz/domain/instrument/repository/InstrumentRepositoryCustomImpl.java
@@ -3,7 +3,7 @@ package com.ajou.hertz.domain.instrument.repository;
 import static com.ajou.hertz.domain.instrument.entity.QAcousticAndClassicGuitar.*;
 import static com.ajou.hertz.domain.instrument.entity.QAmplifier.*;
 import static com.ajou.hertz.domain.instrument.entity.QAudioEquipment.*;
-import static com.ajou.hertz.domain.instrument.entity.QBassGuitar.*;
+import static com.ajou.hertz.domain.instrument.bass_guitar.entity.QBassGuitar.*;
 import static com.ajou.hertz.domain.instrument.entity.QEffector.*;
 import static com.ajou.hertz.domain.instrument.electric_guitar.entity.QElectricGuitar.*;
 import static com.ajou.hertz.domain.user.entity.QUser.*;
@@ -23,14 +23,14 @@ import com.ajou.hertz.domain.instrument.constant.InstrumentSortOption;
 import com.ajou.hertz.domain.instrument.dto.request.AcousticAndClassicGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.AmplifierFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.AudioEquipmentFilterConditions;
-import com.ajou.hertz.domain.instrument.dto.request.BassGuitarFilterConditions;
+import com.ajou.hertz.domain.instrument.bass_guitar.dto.request.BassGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.EffectorFilterConditions;
 import com.ajou.hertz.domain.instrument.electric_guitar.dto.request.ElectricGuitarFilterConditions;
 import com.ajou.hertz.domain.instrument.dto.request.InstrumentFilterConditions;
 import com.ajou.hertz.domain.instrument.entity.AcousticAndClassicGuitar;
 import com.ajou.hertz.domain.instrument.entity.Amplifier;
 import com.ajou.hertz.domain.instrument.entity.AudioEquipment;
-import com.ajou.hertz.domain.instrument.entity.BassGuitar;
+import com.ajou.hertz.domain.instrument.bass_guitar.entity.BassGuitar;
 import com.ajou.hertz.domain.instrument.entity.Effector;
 import com.ajou.hertz.domain.instrument.electric_guitar.entity.ElectricGuitar;
 import com.ajou.hertz.domain.instrument.entity.Instrument;
