+++ src/main/java/com/ajou/hertz/domain/instrument/service/InstrumentCommandService.java
@@ -9,20 +9,20 @@ import org.springframework.web.multipart.MultipartFile;
 import com.ajou.hertz.domain.instrument.dto.AcousticAndClassicGuitarDto;
 import com.ajou.hertz.domain.instrument.dto.AmplifierDto;
 import com.ajou.hertz.domain.instrument.dto.AudioEquipmentDto;
-import com.ajou.hertz.domain.instrument.dto.BassGuitarDto;
+import com.ajou.hertz.domain.instrument.bass_guitar.dto.BassGuitarDto;
 import com.ajou.hertz.domain.instrument.dto.EffectorDto;
 import com.ajou.hertz.domain.instrument.electric_guitar.dto.ElectricGuitarDto;
 import com.ajou.hertz.domain.instrument.dto.request.CreateNewAcousticAndClassicGuitarRequest;
 import com.ajou.hertz.domain.instrument.dto.request.CreateNewAmplifierRequest;
 import com.ajou.hertz.domain.instrument.dto.request.CreateNewAudioEquipmentRequest;
-import com.ajou.hertz.domain.instrument.dto.request.CreateNewBassGuitarRequest;
+import com.ajou.hertz.domain.instrument.bass_guitar.dto.request.CreateNewBassGuitarRequest;
 import com.ajou.hertz.domain.instrument.dto.request.CreateNewEffectorRequest;
 import com.ajou.hertz.domain.instrument.electric_guitar.dto.request.CreateNewElectricGuitarRequest;
 import com.ajou.hertz.domain.instrument.dto.request.CreateNewInstrumentRequest;
 import com.ajou.hertz.domain.instrument.entity.AcousticAndClassicGuitar;
 import com.ajou.hertz.domain.instrument.entity.Amplifier;
 import com.ajou.hertz.domain.instrument.entity.AudioEquipment;
-import com.ajou.hertz.domain.instrument.entity.BassGuitar;
+import com.ajou.hertz.domain.instrument.bass_guitar.entity.BassGuitar;
 import com.ajou.hertz.domain.instrument.entity.Effector;
 import com.ajou.hertz.domain.instrument.electric_guitar.entity.ElectricGuitar;
 import com.ajou.hertz.domain.instrument.entity.Instrument;
@@ -35,7 +35,7 @@ import com.ajou.hertz.domain.instrument.repository.InstrumentRepository;
 import com.ajou.hertz.domain.instrument.strategy.AcousticAndClassicGuitarCreationStrategy;
 import com.ajou.hertz.domain.instrument.strategy.AmplifierCreationStrategy;
 import com.ajou.hertz.domain.instrument.strategy.AudioEquipmentCreationStrategy;
-import com.ajou.hertz.domain.instrument.strategy.BassGuitarCreationStrategy;
+import com.ajou.hertz.domain.instrument.bass_guitar.strategy.BassGuitarCreationStrategy;
 import com.ajou.hertz.domain.instrument.strategy.EffectorCreationStrategy;
 import com.ajou.hertz.domain.instrument.electric_guitar.strategy.ElectricGuitarCreationStrategy;
 import com.ajou.hertz.domain.instrument.strategy.InstrumentCreationStrategy;
