+++ src/test/java/com/ajou/hertz/unit/domain/instrument/service/InstrumentImageCommandServiceTest.java
@@ -21,12 +21,12 @@ import org.springframework.web.multipart.MultipartFile;
 import com.ajou.hertz.common.entity.Address;
 import com.ajou.hertz.common.file.dto.FileDto;
 import com.ajou.hertz.common.file.service.FileService;
-import com.ajou.hertz.domain.instrument.constant.BassGuitarBrand;
-import com.ajou.hertz.domain.instrument.constant.BassGuitarPickUp;
-import com.ajou.hertz.domain.instrument.constant.BassGuitarPreAmplifier;
+import com.ajou.hertz.domain.instrument.bass_guitar.constant.BassGuitarBrand;
+import com.ajou.hertz.domain.instrument.bass_guitar.constant.BassGuitarPickUp;
+import com.ajou.hertz.domain.instrument.bass_guitar.constant.BassGuitarPreAmplifier;
 import com.ajou.hertz.domain.instrument.constant.GuitarColor;
 import com.ajou.hertz.domain.instrument.constant.InstrumentProgressStatus;
-import com.ajou.hertz.domain.instrument.entity.BassGuitar;
+import com.ajou.hertz.domain.instrument.bass_guitar.entity.BassGuitar;
 import com.ajou.hertz.domain.instrument.entity.Instrument;
 import com.ajou.hertz.domain.instrument.entity.InstrumentImage;
 import com.ajou.hertz.domain.instrument.repository.InstrumentImageRepository;
